requirements.txt
flask==3.1.1
requests==2.32.4
pyusb==1.3.1
hidapi==0.14.0.post4
python-dotenv==1.0.0
LabJackPython
